ReversiblePanelの反転処理完了

+++ Project1/Project1/ObjReversibleMain.h
@@ -13,7 +13,8 @@ public:
 	void Init();    //イニシャライズ
 	void Action();  //アクション
 	void Draw();    //ドロー
-	int x, y;
+	float x, y;
+
 private:
 	int stage[5][5];
